feat(images): moved GenericTable.tsx and subcomponents to base/components, fixed th-td alignment for SMImagesTable.tsx
