+++ src/app/base/components/GenericTable/_index.scss
@@ -1,0 +1,11 @@
+.generic-table {
+  table-layout: auto;
+  th:not(th:nth-of-type(1), th:nth-of-type(2), th:last-of-type) {
+    padding-left: 1.5rem;
+  }
+
+  .individual-row {
+    width: calc(100% - 1rem);
+    margin-left: 1rem;
+  }
+}
