--- src/app/images/components/GenericTable/_index.scss
@@ -1,22 +1,0 @@
-.generic-table {
-  th.release {
-    width: 12rem;
-  }
-
-  td.release {
-    width: 11rem;
-  }
-
-  .architecture, .size {
-    width: 9rem;
-  }
-
-  .status {
-    width: 17rem;
-  }
-
-  .individual-row {
-    width: calc(100% - 1rem);
-    margin-left: 1rem;
-  }
-}
