+++ src/app/images/components/SMImagesTable/_index.scss
@@ -1,0 +1,5 @@
+.generic-table {
+  .status {
+    width: 13rem;
+  }
+}
